# Update EdDSA to last circomlib version